+++ src/Tests/Microsoft.DotNet.PackageValidation.Tests/TestLogger.cs
@@ -2,17 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Generic;
-using Microsoft.NET.Build.Tasks;
 
 namespace Microsoft.DotNet.PackageValidation.Tests
 {
-    internal class TestLogger : Logger
+    public class TestLogger : ILogger
     {
         public List<string> errors = new();
 
-        protected override void LogCore(in Message message) 
+        public void LogError(string message)
         {
-            errors.Add(message.Text);
+            errors.Add(message);
         }
     }
 }
